rename undulators file
